Refactors coordinate traits for improved geometry
Consolidates coordinate-related traits into a single module,
enhancing code organization and simplifying trait bounds on
geometric structures.

This change introduces the `Coordinate` trait and supporting
traits like `FiniteCheck`, `OrderedEq`, and `HashCoordinate`
to provide a unified interface for coordinate operations.
The refactor improves NaN handling, enables consistent hashing,
and streamlines validation of coordinate values, ensuring
mathematical correctness and type safety.
